+++ packages/tezos/src/utils/isNetwork.ts
@@ -1,0 +1,11 @@
+import { NetworkType, Networks } from "../types";
+
+export const isNetwork = (str: string | undefined): str is Networks => {
+  if (!str) return false;
+  return Object.values<string>(NetworkType).includes(str.toLowerCase());
+};
+
+export const isTempleNetwork = (str: string | undefined): str is Networks => {
+  if (!str) return false;
+  return Object.values(NetworkType).includes(str.toLowerCase() as Networks);
+};