+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,47 @@
 # example
 
+## 0.1.40-beta.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/tezos@0.2.0-beta.4
+
+## 0.1.40-beta.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/tezos@0.2.0-beta.4
+
+## 0.1.40-beta.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/tezos@0.2.0-beta.3
+
+## 0.1.40-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/tezos@0.2.0-beta.2
+
+## 0.1.40-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/tezos@0.2.0-beta.1
+
+## 0.1.40-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @wallet01/tezos@0.2.0-beta.0
+
 ## 0.1.39
 
 ### Patch Changes