+++ packages/tezos/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @wallet01/tezos
 
+## 0.2.0
+
+### Minor Changes
+
+- 6275204: Adding Beacon connector
+
+### Patch Changes
+
+- edfcb84: changing package imports
+- a0b43cf: Changing network types
+- a0b43cf: changin types locally
+- 8f1a6c2: fixing disconnect function
+- bfe27d6: Adding beacon package in the dependency
+- 8ff3492: Changing import from airgap package
+
+## 0.2.0-beta.6
+
+### Patch Changes
+
+- fixing disconnect function
+
+## 0.2.0-beta.5
+
+### Patch Changes
+
+- changin types locally
+
+## 0.2.0-beta.4
+
+### Patch Changes
+
+- Changing network types
+
+## 0.2.0-beta.3
+
+### Patch Changes
+
+- changing package imports
+
+## 0.2.0-beta.2
+
+### Patch Changes
+
+- Changing import from airgap package
+
+## 0.2.0-beta.1
+
+### Patch Changes
+
+- Adding beacon package in the dependency
+
+## 0.2.0-beta.0
+
+### Minor Changes
+
+- Adding Beacon connector
+
 ## 0.1.5
 
 ### Patch Changes